+++ src/main/java/rs/ac/uns/ftn/siit/team11/ProjectSIIT2023Team11/controller/AvailabilityController.java
@@ -1,13 +1,19 @@
 package rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.actuate.autoconfigure.observation.ObservationProperties;
+import org.springframework.http.HttpOutputMessage;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Accommodation;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Availability;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.Price;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.domain.TimeSlot;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.dto.AccommodationDTO.AvailabilityDTO;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.mapper.AvailabilityMapper;
+import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.service.IAccommodationService;
 import rs.ac.uns.ftn.siit.team11.ProjectSIIT2023Team11.service.IAvailabilityService;
 
 import java.util.Collection;
@@ -21,6 +27,8 @@ public class AvailabilityController {
 
     @Autowired
     private IAvailabilityService availabilityService;
+    @Autowired
+    private IAccommodationService accommodationService;
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity<Collection<AvailabilityDTO>> getAllAvailabilities() {
@@ -49,10 +57,50 @@ public class AvailabilityController {
     }
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<Void> deleteAvailability(@PathVariable("id") Long id) {
+        Optional<Availability> availability = availabilityService.findById(id);
+        if(availability.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        accommodationService.deleteAvailabilityFromAllAccommodations(availability.get());
         availabilityService.deleteById(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
+    @PostMapping( value = "/accommodation/{accommodation_id}",
+            consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Availability> createAvailability(@RequestBody AvailabilityDTO availability, @PathVariable("accommodation_id") Long accommodationId) {
+        Optional<Accommodation> accommodation = accommodationService.findById(accommodationId);
+        if(accommodation.isEmpty())
+        {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        if(accommodation.get().AddAvailability(new TimeSlot(availability.getStartDate().plusDays(1), availability.getEndDate().plusDays(1)))) return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        Availability createdAvailability = availabilityService.save(AvailabilityMapper.mapToAvailability(availability));
 
+        accommodation.get().getAvailability().add(createdAvailability);
+        accommodationService.save(accommodation.get());
+
+        return new ResponseEntity<>(createdAvailability, HttpStatus.CREATED);
+    }
+
+    @DeleteMapping(value = "/{id}/accommodation/{accommodationId}")
+    public ResponseEntity<Void> deleteAccommodationAvailability(@PathVariable("id") Long id, @PathVariable("accommodationId") Long accommodationId) {
+
+        Optional<Accommodation> accommodation = accommodationService.findById(accommodationId);
+        Optional<Availability> availability = availabilityService.findById(id);
+        if(accommodation.isEmpty() || availability.isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        while(accommodation.get().getAvailability().contains(availability.get())){
+            accommodation.get().getAvailability().remove(availability.get());
+
+        }
+
+        availabilityService.deleteById(id);
+
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+
+    }
 
 }
